Modifique el metodo de modulo, para que pueda mostrar decimales

+++ src/operadores.java
@@ -28,8 +28,8 @@ public class operadores {
         System.out.println("El resultado de la division es:" + resultado);
         return resultado;
     }
-    public int modulo (int op1, int op2){
-        int resultado=0;
+    public float modulo (int op1, int op2){
+        float resultado=0;
 
         resultado = op1 % op2;
         System.out.println("El resultado de la modulo es:" + resultado);
